ci: Enhance Deployment Process with Health Check & Improved Error Handling

+++ server/tests/integration/routes/auth.test.ts
@@ -1,17 +1,19 @@
 import bcrypt from "bcrypt";
 import cookie from "cookie";
 import "dotenv/config";
-import http from "http";
+import { Server } from "http";
+import { Application } from "express";
 import { disconnect } from "mongoose";
+import ms from "ms";
 import request from "supertest";
 
-import ms from "ms";
-import appServer from "../../../src";
 import { IUser, User } from "../../../src/models/user.model";
 import FirebaseAuthError from "../../../src/utils/errors/firebase-error";
 import * as firebaseAuth from "../../../src/utils/firebase-auth";
+import { startServer } from "../../../src/start";
 
-let server: http.Server;
+let server: Server;
+let app: Application;
 let endpoint: string = `/api/v1/auth`;
 
 type DecodedIdToken = {
@@ -34,18 +36,23 @@ type DecodedIdToken = {
 };
 
 describe("/api/v1/auth", () => {
+  beforeAll(async () => {
+    try {
+      ({ server, app } = await startServer());
+    } catch (error) {
+      console.error("🚨 Server startup failed in tests:", error);
+      throw new Error("Failed to start the test server");
+    }
+  });
+
   afterAll(async () => {
-    // close the MongoDB connection
+    if (!server) return;
+    server.close();
     await disconnect();
   });
 
-  beforeEach(() => {
-    server = appServer;
-  });
-
   afterEach(async () => {
-    server.close();
-    // db cleanup
+    if (!server) return;
     await User.deleteMany({});
   });
 
@@ -55,7 +62,7 @@ describe("/api/v1/auth", () => {
       const userData = {
         password: "clubhouse",
       };
-      const res = await request(server).post(endpoint).send(userData);
+      const res = await request(app).post(endpoint).send(userData);
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -71,7 +78,7 @@ describe("/api/v1/auth", () => {
         email: "test@test.com",
         password: "clubhouse",
       };
-      const res = await request(server).post(endpoint).send(userData);
+      const res = await request(app).post(endpoint).send(userData);
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -95,7 +102,7 @@ describe("/api/v1/auth", () => {
         email: "test@test.com",
         password: "Pluto123",
       };
-      const res = await request(server).post(endpoint).send(userData);
+      const res = await request(app).post(endpoint).send(userData);
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -121,7 +128,7 @@ describe("/api/v1/auth", () => {
         email: "test@test.com",
         password: "Clubhouse123",
       };
-      const res = await request(server).post(endpoint).send(userData);
+      const res = await request(app).post(endpoint).send(userData);
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -148,7 +155,7 @@ describe("/api/v1/auth", () => {
         email: "test@gmail.com",
         password: "Clubhouse123",
       };
-      const res = await request(server).post(endpoint).send(userData);
+      const res = await request(app).post(endpoint).send(userData);
 
       expect(res.statusCode).toBe(403);
 
@@ -162,7 +169,7 @@ describe("/api/v1/auth", () => {
 
     it("should authenticate user if request is valid", async () => {
       // call /register route so it can create user and store hash password
-      const registerRes = await request(server)
+      const registerRes = await request(app)
         .post(`/api/v1/users/register`)
         .send({
           fullname: "Mickey Mouse",
@@ -184,7 +191,7 @@ describe("/api/v1/auth", () => {
         email: "test@test.com",
         password: "Pluto123",
       };
-      const res = await request(server).post(endpoint).send(userData);
+      const res = await request(app).post(endpoint).send(userData);
 
       expect(res.statusCode).toBe(200);
       expect(res.body.status).toBe("success");
@@ -205,7 +212,7 @@ describe("/api/v1/auth", () => {
 
   describe("POST /google-auth", () => {
     const exec = async (accessToken: string) => {
-      return await request(server).post(`${endpoint}/google-auth`).send({
+      return await request(app).post(`${endpoint}/google-auth`).send({
         accessToken,
       });
     };
@@ -220,7 +227,7 @@ describe("/api/v1/auth", () => {
       const userData = {
         token: "",
       };
-      const res = await request(server)
+      const res = await request(app)
         .post(`${endpoint}/google-auth`)
         .send(userData);
       expect(res.statusCode).toBe(400);
@@ -369,7 +376,7 @@ describe("/api/v1/auth", () => {
     let token: string;
 
     const exec = async () => {
-      return await request(server)
+      return await request(app)
         .post(`${endpoint}/logout`)
         .set("Cookie", `authToken=${token}`);
     };
@@ -408,6 +415,7 @@ describe("/api/v1/auth", () => {
 
   describe("POST /verify-email", () => {
     afterEach(async () => {
+      if (!server) return;
       // db cleanup
       await User.deleteMany({});
     });
@@ -415,7 +423,7 @@ describe("/api/v1/auth", () => {
     it("should return BadRequest-400 if token is not passed", async () => {
       // email and token are the required parameter for token verification.
       const email = "test@test.com";
-      const res = await request(server).post(`${endpoint}/verify-email`).send({
+      const res = await request(app).post(`${endpoint}/verify-email`).send({
         email,
       });
 
@@ -431,7 +439,7 @@ describe("/api/v1/auth", () => {
       // token is invalid
       const email = "test@test.com";
       const token = "invalid-token";
-      const res = await request(server).post(`${endpoint}/verify-email`).send({
+      const res = await request(app).post(`${endpoint}/verify-email`).send({
         email,
         token,
       });
@@ -463,7 +471,7 @@ describe("/api/v1/auth", () => {
       await user.save();
 
       const email = user.personalInfo.email;
-      const res = await request(server).post(`${endpoint}/verify-email`).send({
+      const res = await request(app).post(`${endpoint}/verify-email`).send({
         email,
         token,
       });
@@ -489,7 +497,7 @@ describe("/api/v1/auth", () => {
 
       const token = "some-random-token";
       const email = user.personalInfo.email;
-      const res = await request(server).post(`${endpoint}/verify-email`).send({
+      const res = await request(app).post(`${endpoint}/verify-email`).send({
         email,
         token,
       });
@@ -517,7 +525,7 @@ describe("/api/v1/auth", () => {
       expect(user.isVerified).toBeFalsy();
 
       const email = user.personalInfo.email;
-      const res = await request(server).post(`${endpoint}/verify-email`).send({
+      const res = await request(app).post(`${endpoint}/verify-email`).send({
         email,
         token,
       });
@@ -539,13 +547,14 @@ describe("/api/v1/auth", () => {
 
   describe("POST /resend-verification", () => {
     afterEach(async () => {
+      if (!server) return;
       // db cleanup
       await User.deleteMany({});
     });
 
     it("should avoid explicitly indicating if the email is unregistered", async () => {
       const email = "randomuser@test.com";
-      const res = await request(server)
+      const res = await request(app)
         .post(`${endpoint}/resend-verification`)
         .send({ email });
 
@@ -566,7 +575,7 @@ describe("/api/v1/auth", () => {
         isVerified: true,
       });
 
-      const res = await request(server)
+      const res = await request(app)
         .post(`${endpoint}/resend-verification`)
         .send({
           email: user.personalInfo.email,
@@ -598,7 +607,7 @@ describe("/api/v1/auth", () => {
       };
       await user.save();
 
-      const res = await request(server)
+      const res = await request(app)
         .post(`${endpoint}/resend-verification`)
         .send({
           email: user.personalInfo.email,
@@ -630,7 +639,7 @@ describe("/api/v1/auth", () => {
       };
       await user.save();
 
-      const res = await request(server)
+      const res = await request(app)
         .post(`${endpoint}/resend-verification`)
         .send({
           email: user.personalInfo.email,
@@ -645,13 +654,14 @@ describe("/api/v1/auth", () => {
 
   describe("POST /forgot-password", () => {
     afterEach(async () => {
+      if (!server) return;
       // db cleanup
       await User.deleteMany({});
     });
 
     it("should avoid explicitly indicating if the email is unregistered", async () => {
       const email = "randomuser@test.com";
-      const res = await request(server)
+      const res = await request(app)
         .post(`${endpoint}/forgot-password`)
         .send({ email });
 
@@ -679,11 +689,9 @@ describe("/api/v1/auth", () => {
       };
       await user.save();
 
-      const res = await request(server)
-        .post(`${endpoint}/forgot-password`)
-        .send({
-          email: user.personalInfo.email,
-        });
+      const res = await request(app).post(`${endpoint}/forgot-password`).send({
+        email: user.personalInfo.email,
+      });
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -711,11 +719,9 @@ describe("/api/v1/auth", () => {
       };
       await user.save();
 
-      const res = await request(server)
-        .post(`${endpoint}/forgot-password`)
-        .send({
-          email: user.personalInfo.email,
-        });
+      const res = await request(app).post(`${endpoint}/forgot-password`).send({
+        email: user.personalInfo.email,
+      });
 
       expect(res.statusCode).toBe(200);
       expect(res.body.message).toMatch(
@@ -726,6 +732,7 @@ describe("/api/v1/auth", () => {
 
   describe("POST /reset-password", () => {
     afterEach(async () => {
+      if (!server) return;
       // db cleanup
       await User.deleteMany({});
     });
@@ -733,11 +740,9 @@ describe("/api/v1/auth", () => {
     it("should return BadRequest-400 if token is not passed", async () => {
       // token is the required parameter for token verification.
       const email = "test@test.com";
-      const res = await request(server)
-        .post(`${endpoint}/reset-password`)
-        .send({
-          email,
-        });
+      const res = await request(app).post(`${endpoint}/reset-password`).send({
+        email,
+      });
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -752,13 +757,11 @@ describe("/api/v1/auth", () => {
       const email = "test@test.com";
       const token = "random-token";
       const newPassword = "invalid-password";
-      const res = await request(server)
-        .post(`${endpoint}/reset-password`)
-        .send({
-          email,
-          token,
-          password: newPassword,
-        });
+      const res = await request(app).post(`${endpoint}/reset-password`).send({
+        email,
+        token,
+        password: newPassword,
+      });
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -774,13 +777,11 @@ describe("/api/v1/auth", () => {
       const email = "test@test.com";
       const token = "invalid-token";
       const newPassword = "Clubhouse123";
-      const res = await request(server)
-        .post(`${endpoint}/reset-password`)
-        .send({
-          email,
-          token,
-          password: newPassword,
-        });
+      const res = await request(app).post(`${endpoint}/reset-password`).send({
+        email,
+        token,
+        password: newPassword,
+      });
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -809,13 +810,11 @@ describe("/api/v1/auth", () => {
       await user.save();
 
       const newPassword = "Pluto123";
-      const res = await request(server)
-        .post(`${endpoint}/reset-password`)
-        .send({
-          email: user.personalInfo.email,
-          token,
-          password: newPassword,
-        });
+      const res = await request(app).post(`${endpoint}/reset-password`).send({
+        email: user.personalInfo.email,
+        token,
+        password: newPassword,
+      });
 
       expect(res.statusCode).toBe(400);
       expect(res.body.error).toMatchObject({
@@ -843,13 +842,11 @@ describe("/api/v1/auth", () => {
       await user.save();
       const newPassword = "Pluto123";
 
-      const res = await request(server)
-        .post(`${endpoint}/reset-password`)
-        .send({
-          email: user.personalInfo.email,
-          token,
-          password: newPassword,
-        });
+      const res = await request(app).post(`${endpoint}/reset-password`).send({
+        email: user.personalInfo.email,
+        token,
+        password: newPassword,
+      });
 
       expect(res.statusCode).toBe(200);
       expect(res.body.message).toBe("Password reset completed successfully.");
